backend/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework>=3.9,<4.0
markdown>=3.1,<4.0
django-filter>=2.1,<3.0
django-mptt>=0.10.0